+++ module-cellular/at/src/URC_CUSD.cpp
@@ -2,32 +2,79 @@
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include <URC_CUSD.hpp>
+#include <Utils.hpp>
+#include <magic_enum.hpp>
 
 using namespace at::urc;
 
-CUSD::CUSD(const std::string &val) : Any(val, '\"')
+CUSD::CUSD(const std::string &val) : Any(val)
 {}
 
-auto CUSD::what() -> std::string
+auto CUSD::what() const noexcept -> std::string
 {
     return urc_name;
 }
 
-bool CUSD::isActionNeeded(void)
+auto CUSD::isValid() const noexcept -> bool
 {
-    constexpr uint32_t actionNeeded = 2;
+    return is() && tokens.size() == magic_enum::enum_count<Tokens>();
+}
 
-    utils::findAndReplaceAll(tokens[static_cast<uint32_t>(CUSD::Tokens::Status)], ",", "");
-    uint32_t status = 0;
-    if (utils::toNumeric(tokens[static_cast<uint32_t>(CUSD::Tokens::Status)], status)) {
-        if (status == actionNeeded) {
-            return true;
+auto CUSD::isActionNeeded() const noexcept -> bool
+{
+    if (isValid()) {
+        auto status = getStatus();
+        if (status) {
+            return *status == StatusType::FurtherUserActionRequired;
         }
     }
     return false;
 }
 
-std::string CUSD::message(void)
+auto CUSD::getMessage() const noexcept -> std::optional<std::string>
 {
-    return tokens[static_cast<uint32_t>(CUSD::Tokens::Response)];
+    if (!isValid()) {
+        return std::nullopt;
+    }
+
+    auto message = tokens[Tokens::Response];
+    utils::findAndReplaceAll(message, "\"", "");
+
+    return utils::trim(message);
+}
+
+auto CUSD::getStatus() const noexcept -> std::optional<StatusType>
+{
+    if (isValid()) {
+        int statusInt;
+        try {
+            statusInt = std::stoi(tokens[Tokens::Status]);
+        }
+        catch (const std::exception &e) {
+            return std::nullopt;
+        }
+
+        auto status = magic_enum::enum_cast<StatusType>(statusInt);
+        if (status.has_value()) {
+            return status.value();
+        }
+    }
+
+    return std::nullopt;
+}
+
+auto CUSD::getDCS() const noexcept -> std::optional<int>
+{
+    if (!isValid()) {
+        return std::nullopt;
+    }
+    int dcs;
+    try {
+        dcs = std::stoi(tokens[Tokens::DCS]);
+    }
+    catch (const std::exception &e) {
+        return std::nullopt;
+    }
+
+    return dcs;
 }